refactor(RestaurantIcon): RestaurantIcon을 TS로 마이그레이션
